Integrate with provider configuration
